Token login failure fixed
This commit fixes the token login.
The issue was that when logging in via username and password the user
would get sent back a token with a completely different secret than the
one being checked via logging in with token, so the quick fix was just
getting Logging in via username and password to use the one the token
login uses.